+++ packages/react-native-sdk/expo-config-plugin/__tests__/withAndroidManifest.test.ts
@@ -38,9 +38,7 @@ const props: ConfigProps = {
   ringingPushNotifications: {
     disableVideoIos: false,
   },
-  androidPictureInPicture: {
-    enableAutomaticEnter: true,
-  },
+  androidPictureInPicture: true,
   androidKeepCallAlive: true,
 };